requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
pydantic-settings
python-dotenv
icalendar
pytz
notion-client
python-dateutil
